15.09.2020

+++ 6.BancoDeDados/5.CRUD-DELETE/index.php
@@ -26,7 +26,7 @@ if($sql->rowCount()>0){//Verifica se tem algum
             <td><?= $usuario['email']; ?></td>
             <td>
                 <a href="editar.php?id=<?= $usuario['id']; ?>">[ Editar ]</a>
-                <a href="excluir.php?id=<?= $usuario['id']; ?>">[ Excluir ]</a>
+                <a href="excluir.php?id=<?= $usuario['id']; ?> " onclick="return confirm('Tem certeza?')">[ Excluir ]</a>
             </td>
         </tr>
     <?php endforeach; ?>
